refactor: 메인페이지 서버컴포 전환 및 버튼 클라컴포로 분리 (#28)
